Added Game States, Pickups, Timer, Labels
Redesign for this Game:

Pick up 15 Sandler Spheres within an a set time frame. If the time limit
is reached without picking up all of the Sandlers, you lose; if you get
15 Sandlers within the time limit, you win.

What's left to do:

Add Audio, set a certain time limit that is acceptable for a challenge,
have a score controller and time controller to bring the number of the
acquired pickups to Game Over state, and the best time run to the Win
state.